Move all patching to client-side
