+++ clients/assets/bootloader/index.html
@@ -1,0 +1,10 @@
+<!DOCTYPE html>
+<html>
+	<head>
+		<meta charset="utf-8">
+		<meta content="width=device-width, initial-scale=1.0, maximum-scale=1, user-scalable=no" name="viewport">
+		<link rel="icon" href="data:;base64,iVBORw0KGgo=">
+		<title>Oldcord</title>
+		<script src="/assets/bootloader/bootloader.js"></script>
+	</head>
+</html>
